moving HQ to proper location
